requirements.txt
requests==2.5.2
tapioca-wrapper==0.2.0
